+++ tests/utils/zeek_sender.py
@@ -5,45 +5,47 @@ import select
 import time
 
 
-def send(topic, broker_event):
-    """Sends a single, user specified broker event"""
+def peer(host, port):
+    """
+    Peers via Broker with the Threat Bus Zeek plugin
+    """
     ep = broker.Endpoint()
     status_subscriber = ep.make_status_subscriber(True)
-    ep.peer("127.0.0.1", 47761)
+    ep.peer(host, port)
 
     fd_sets = select.select([status_subscriber.fd()], [], [])
     if not fd_sets[0]:
         print("Peering with remote machine failed")
-        return
+        return False
     status = status_subscriber.get()
     if type(status) != broker.Status or status.code() != broker.SC.PeerAdded:
         print("Threat Bus subscription failed")
-        return
+        return False
+    return ep
+
+
+def send(topic, broker_event):
+    """Sends a single, user specified broker event"""
+    ep = peer("127.0.0.1", 47761)
+    if not ep:
+        return False
 
     ep.publish(topic, broker_event)
     time.sleep(0.1)
 
 
 def send_generic(topic, items):
-    ep = broker.Endpoint()
-    status_subscriber = ep.make_status_subscriber(True)
-    ep.peer("127.0.0.1", 47761)
+    ep = peer("127.0.0.1", 47761)
+    if not ep:
+        return False
 
-    fd_sets = select.select([status_subscriber.fd()], [], [])
-    if not fd_sets[0]:
-        print("Peering with remote machine failed")
-        return
-    status = status_subscriber.get()
-    if type(status) != broker.Status or status.code() != broker.SC.PeerAdded:
-        print("Threat Bus subscription failed")
-        return
-
-    for i in range(items):
-        data = {
-            "indicator": "example.com",
-            "intel_type": "DOMAIN",
-        }
-        event = broker.zeek.Event("intel", datetime.now(), i, data, "ADD")
+    for _ in range(items):
+        event = broker.zeek.Event(
+            "sighting",
+            datetime.now(),
+            "indicator--cdd5791f-916e-4f62-8090-1a006005af76",
+            {},
+        )
 
         # Threat Bus will pickup the event type and hence forward on a different
         # topic.
@@ -54,4 +56,4 @@ def send_generic(topic, items):
 
 
 if __name__ == "__main__":
-    send_generic("threatbus/intel", 1)
+    send_generic("stix2/sighting", 1)